# Support git partial clones with sparse checkouts