--- a/Library/Homebrew/download_strategy.rb
+++ b/Library/Homebrew/download_strategy.rb
@@ -927,12 +927,12 @@ class GitDownloadStrategy < VCSDownloadStrategy
              args:  ["config", "advice.detachedHead", "false"],
              chdir: cached_location
 
-    if partial_clone_sparse_checkout?
-      command! "git",
-               args:  ["config", "origin.partialclonefilter", "blob:none"],
-               chdir: cached_location
-      configure_sparse_checkout
-    end
+    return unless partial_clone_sparse_checkout?
+
+    command! "git",
+             args:  ["config", "origin.partialclonefilter", "blob:none"],
+             chdir: cached_location
+    configure_sparse_checkout
   end
 
   sig { params(timeout: T.nilable(Time)).void }
@@ -1040,7 +1040,7 @@ class GitDownloadStrategy < VCSDownloadStrategy
              args:  ["config", "core.sparseCheckout", "true"],
              chdir: cached_location
 
-    sparse_checkout_paths = @only_paths.join("\n") + "\n"
+    sparse_checkout_paths = "#{@only_paths.join("\n")}\n"
     (git_dir/"info"/"sparse-checkout").atomic_write(sparse_checkout_paths)
   end
 end