@hmarr 2022-05-04
Thanks for the review @MikeMcQuaid 🙌 

Here's a draft PR showing how the new DSL could be used: https://github.com/Homebrew/homebrew-cask-fonts/pull/5758