@hmarr 2022-05-07
Thanks @MikeMcQuaid!

Before merging, we should probably also decide on a way of handling upgrades from sparse to full checkouts (which would happen in the rare case someone _removes_ `only_paths`). I listed the options I can see in the PR body, but they boil down to:
1. Add something to the cache key to separate sparse checkouts from full checkouts. This is the simpler option to implement, but means that if you remove `only_paths` you'd re-clone the repo from scratch.
2. Detect sparse checkouts and explicitly perform the upgrade when `only_paths` isn't specified. This would let us reuse the sparse checkout when upgrading to a full checkout, but I'm not convinced that marginal speed benefit in what must be a very rare case is worth the extra complexity in the download strategy code.