@MikeMcQuaid 2022-05-15
> * Add something to the cache key to separate sparse checkouts from full checkouts. This is the simpler option to implement, but means that if you remove `only_paths` you'd re-clone the repo from scratch.

This seems fine 👍🏻