@hmarr 2022-06-12
@apainintheneck thanks for the heads up, I made that change in https://github.com/Homebrew/brew/pull/13232/commits/ccdd45dbb7322e60cdb72435e4c8c7562a0e7a8b

@MikeMcQuaid I made the change to the cache path for sparse checkouts in https://github.com/Homebrew/brew/pull/13232/commits/a8c9d3828ce74124341429d51b0be916fada8ea4. That was the last thing I had on my list, so I we're ready to go now. Once this is merged, we can start switching the font casks over to git from svn.