@apainintheneck 2022-06-14
The url documentation over at the [Cask Cookbook](https://docs.brew.sh/Cask-Cookbook#stanza-url) should also probably be updated at some point too.